requirements.txt
requests
beautifulsoup4
lxml
python-dateutil
tqdm
pandas
yt-dlp
pyrate-limiter
PyYAML
